requirements.txt
anyio==4.10.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
h11==0.16.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jose==1.0.0
msgpack==1.1.1
nanoid==2.0.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
requests==2.32.4
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
zope.interface==7.2
